+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Move header reference out of "extern C" (#3538)
 - Clarify FramesTracker log message (#3570)
+- Fix rare battery breadcrumbs crash (#3582)
+
 
 ## 8.19.0